apps/schema-registry-ui/requirements.txt
Flask==3.0.3
PyYAML==6.0.2
jsonschema==4.23.0
